starter cards refactoring
