+++ card/stories/index.stories.tsx
@@ -1,4 +1,4 @@
-import EmptyCard from '../src/EmptyCard.vue';
+import Card from '../src/Card.vue';
 import '~/tailwind';
 
 export default {
@@ -8,12 +8,12 @@ export default {
 };
 
 export const card = (args) => ({
-  components: { EmptyCard },
+  components: { Card },
   template: /*html*/ `
-    <EmptyCard url="/request-earlyaccess" class="bg-primary">
+    <Card url="/request-earlyaccess" class="bg-primary">
       <div class="w-full h-full flex items-center text-3xl sm:text-4xl text-bold text-center text-black-default">
         Many other Starter-kits available in different technologies
       </div>
-    </EmptyCard>
+    </Card>
   `,
 });
