+++ starter-card/src/StarterCard.vue
@@ -1,27 +1,39 @@
 <template>
   <component
-    :is="EmptyCard"
+    :is="Card"
     :url="url"
     @click="(event) => $emit('click', event)"
     :backgroundColor="backgroundColor"
     :size="size"
   >
-    <component :is="StarterLogo" :name="name" :heroImg="heroImg" :size="size" />
+    <div :class="`logo logo-${size}`">
+      <img :src="heroImg" />
+    </div>
+    <header class="starter-title">Starter {{ name }}</header>
     <p class="description">
       <slot></slot>
     </p>
-    <component
-      :is="ActionDescription"
-      :description="actionDescription"
-      :size="size"
-    />
+    <div :class="`action-description action-description-${size}`">
+      {{ actionDescription }}
+      <svg
+        xmlns="http://www.w3.org/2000/svg"
+        style="width: 1.5em; height: 1.5em"
+        aria-hidden="true"
+        viewBox="0 0 20 20"
+        fill="currentColor"
+      >
+        <path
+          fill-rule="evenodd"
+          d="M7.293 14.707a1 1 0 010-1.414L10.586 10 7.293 6.707a1 1 0 011.414-1.414l4 4a1 1 0 010 1.414l-4 4a1 1 0 01-1.414 0z"
+          clip-rule="evenodd"
+        />
+      </svg>
+    </div>
     <img class="framework-bg" :src="bgImg" />
   </component>
 </template>
 <script>
-import EmptyCard from '../../empty-card/src/EmptyCard.vue';
-import ActionDescription from '../../action-description/src/ActionDescription.vue';
-import StarterLogo from '../../starter-logo/src/StarterLogo.vue';
+import Card from '../../card/src/Card.vue';
 
 export default {
   props: {
@@ -34,10 +46,34 @@ export default {
     actionDescription: { type: String },
     size: { type: String, default: 'normal' },
   },
-  data: () => ({ EmptyCard, ActionDescription, StarterLogo }),
+  data: () => ({ Card }),
 };
 </script>
 <style lang="scss" scoped>
+.logo-normal {
+  height: 5rem;
+}
+
+.logo-big {
+  height: 10rem;
+}
+
+.logo {
+  width: 100%;
+  display: flex;
+  justify-content: center;
+  margin-bottom: 1rem;
+  img {
+    max-width: 100%;
+    height: 100%;
+  }
+}
+
+.starter-title {
+  font-size: 1.5em;
+  font-weight: bold;
+}
+
 .description {
   flex-grow: 1;
 }
@@ -50,5 +86,16 @@ export default {
   bottom: 0;
   opacity: 0.25;
   transform: translateX(4rem) translateY(3rem) rotate(-12deg);
+}
+
+.action-description {
+  display: flex;
+  align-items: center;
+  font-weight: bold;
+  z-index: 1;
+}
+
+.action-description-big {
+  font-size: 1.5em;
 }
 </style>
